fix: align telemetry cadences with the CSV writer
- get_power: drop the 5-min gate entirely. The plug is now read every
  step (every minute), matching get_image's cadence. CSV still writes
  every 5 min and just picks up the most recent reading; WandB charts
  gain per-minute granularity. The 5-min throttle was artificial when
  we already pay per-minute network latency for images.
- get_plant_stats: switch from relative `now - last_cv_time >= 5min`
  to the same `(minute % 5 == 0) OR overdue` pattern PR #320 introduced
  for kasa, so the CV fetch lands on the CSV write boundary instead of
  drifting 1-4 minutes ahead.

Drops the now-unused last_smart_plug_time and smart_plug_interval state.
Adds focused unit tests for the get_plant_stats boundary + drift-recovery
behavior; rewrites the smart-plug gate tests as a single
"fetches every call" test now that the kasa gate is gone.
