+++ tests/environments/PlantGrowthChamber/test_SmartPlug_integration.py
@@ -1,12 +1,8 @@
-from datetime import datetime
 from unittest.mock import AsyncMock
-from zoneinfo import ZoneInfo
 
 import pytest
 
 from environments.PlantGrowthChamber.PlantGrowthChamber import PlantGrowthChamber
-
-UTC = ZoneInfo("Etc/UTC")
 
 
 @pytest.fixture
@@ -15,17 +11,10 @@ def kasa_creds(monkeypatch):
     monkeypatch.setenv("KASA_PASSWORD", "p")
 
 
-def _pin_time(chamber, monkeypatch, minute: int):
-    monkeypatch.setattr(
-        chamber, "get_time", lambda: datetime(2026, 5, 6, 12, minute, tzinfo=UTC)
-    )
-
-
 @pytest.mark.asyncio
-async def test_get_power_carryover_records_null_on_failure(kasa_creds, monkeypatch):
+async def test_get_power_carryover_records_null_on_failure(kasa_creds):
     chamber = PlantGrowthChamber(zone="alliance-zone01", timezone="Etc/UTC")
     assert chamber.zone.smart_plug_host == "142.244.4.73"
-    _pin_time(chamber, monkeypatch, minute=5)
 
     successful = {"power": 5.0, "voltage": 120.0, "current": 0.04}
     chamber.smart_plug_client.read = AsyncMock(side_effect=[successful, None])
@@ -63,39 +52,12 @@ async def test_get_power_is_noop_when_zone_has_no_plug(kasa_creds):
 
 
 @pytest.mark.asyncio
-async def test_get_power_fetches_on_five_minute_boundary(kasa_creds, monkeypatch):
+async def test_get_power_fetches_every_call(kasa_creds):
     chamber = PlantGrowthChamber(zone="alliance-zone01", timezone="Etc/UTC")
     successful = {"power": 5.0, "voltage": 120.0, "current": 0.04}
     chamber.smart_plug_client.read = AsyncMock(return_value=successful)
 
-    _pin_time(chamber, monkeypatch, minute=5)
-    await chamber.get_power()
-    assert chamber.power_record == successful
+    for _ in range(3):
+        await chamber.get_power()
 
-    _pin_time(chamber, monkeypatch, minute=6)
-    await chamber.get_power()
-    assert chamber.smart_plug_client.read.await_count == 1
-    assert chamber.power_record == successful
-    assert chamber.power == successful
-
-    _pin_time(chamber, monkeypatch, minute=10)
-    await chamber.get_power()
-    assert chamber.smart_plug_client.read.await_count == 2
-
-
-@pytest.mark.asyncio
-async def test_get_power_recovers_when_boundary_is_missed(kasa_creds, monkeypatch):
-    """If the step loop drifts past a 5-min boundary, the next call still fetches."""
-    chamber = PlantGrowthChamber(zone="alliance-zone01", timezone="Etc/UTC")
-    successful = {"power": 5.0, "voltage": 120.0, "current": 0.04}
-    chamber.smart_plug_client.read = AsyncMock(return_value=successful)
-
-    _pin_time(chamber, monkeypatch, minute=5)
-    await chamber.get_power()
-    assert chamber.smart_plug_client.read.await_count == 1
-
-    # Step loop missed minute=10 entirely, lands at :11. Not on boundary,
-    # but elapsed is 6 min > 5 min interval → fetch via the overdue fallback.
-    _pin_time(chamber, monkeypatch, minute=11)
-    await chamber.get_power()
-    assert chamber.smart_plug_client.read.await_count == 2
+    assert chamber.smart_plug_client.read.await_count == 3
